refactor(backend): split runtime and system layer
