+++ backend/presentation/routes/system.py
@@ -1,24 +1,52 @@
-import asyncio
 import re
 import secrets
-import time
 from typing import Annotated
 
 from fastapi import APIRouter, Depends, HTTPException, Query, Request, Response
 from fastapi.responses import PlainTextResponse
 
 from backend.config.settings import is_valid_admin_token, reload_settings, settings
-from backend.infrastructure.redis_client import redis_cache
 from backend.server.dependencies import get_nesh_service
 from backend.server.middleware import decode_clerk_jwt
 from backend.server.rate_limit import status_rate_limiter
 from backend.services import NeshService
 from backend.utils.auth import extract_bearer_token, extract_client_ip, is_admin_payload
 
+from .system_metrics import append_catalog_status_metrics
+from .system_metrics import append_database_latency_metric
+from .system_metrics import append_internal_cache_hit_rate_metrics
+from .system_metrics import append_metric_line
+from .system_metrics import append_payload_cache_metrics
+from .system_metrics import build_prometheus_metrics_payload
+from .system_metrics import metric_value_from_status
+from .system_status import await_status_refresh_snapshot
+from .system_status import build_detailed_status_payload
+from .system_status import build_public_status_payload
+from .system_status import build_status_snapshot
+from .system_status import coerce_int
+from .system_status import collect_db_status
+from .system_status import collect_nbs_catalog_health
+from .system_status import collect_status_payloads
+from .system_status import collect_status_payloads_uncached
+from .system_status import collect_tipi_status
+from .system_status import extract_prefixed_metadata
+from .system_status import get_status_cache_lock
+from .system_status import get_status_snapshot
+from .system_status import normalize_count_catalog_status
+from .system_status import normalize_db_status
+from .system_status import normalize_tipi_status
+from .system_status import read_l1_status_snapshot
+from .system_status import read_redis_status_snapshot
+from .system_status import read_stale_l1_status_snapshot
+from .system_status import recover_status_snapshot
+from .system_status import refresh_status_snapshot
+from .system_status import reset_status_cache_for_tests
+from .system_status import status_cache_ttl_seconds
+from .system_status import store_status_snapshot
+from .system_status import unpack_status_snapshot
+
 router = APIRouter()
-_STATUS_CACHE: dict[str, object | None] = {"value": None, "expires_at": 0.0}
-_STATUS_CACHE_REFRESH_TASK: asyncio.Task | None = None
-_STATUS_CACHE_LOCK: asyncio.Lock | None = None
+collect_nbs_status = collect_nbs_catalog_health
 STATUS_RESPONSES = {
     429: {"description": "Limite de requisições para status excedido."},
 }
@@ -30,41 +58,6 @@ METRICS_RESPONSES = {
     403: {"description": "Forbidden (invalid metrics token)."},
     404: {"description": "Not Found when metrics endpoint is disabled."},
 }
-
-
-def _get_status_cache_lock() -> asyncio.Lock:
-    global _STATUS_CACHE_LOCK
-    if _STATUS_CACHE_LOCK is None:
-        _STATUS_CACHE_LOCK = asyncio.Lock()
-    return _STATUS_CACHE_LOCK
-
-
-def _status_cache_ttl_seconds() -> int:
-    return max(int(getattr(settings.cache, "status_cache_ttl", 0) or 0), 0)
-
-
-def _read_l1_status_snapshot(now: float | None = None) -> dict | None:
-    snapshot = _STATUS_CACHE.get("value")
-    if not isinstance(snapshot, dict):
-        return None
-    if now is None:
-        now = time.monotonic()
-    expires_at = float(_STATUS_CACHE.get("expires_at") or 0.0)
-    if expires_at > now:
-        return snapshot
-    return None
-
-
-def _read_stale_l1_status_snapshot() -> dict | None:
-    snapshot = _STATUS_CACHE.get("value")
-    return snapshot if isinstance(snapshot, dict) else None
-
-
-def _reset_status_cache_for_tests() -> None:
-    global _STATUS_CACHE_REFRESH_TASK
-    _STATUS_CACHE["value"] = None
-    _STATUS_CACHE["expires_at"] = 0.0
-    _STATUS_CACHE_REFRESH_TASK = None
 
 
 async def _is_admin_request(request: Request) -> bool:
@@ -87,7 +80,6 @@ def _extract_metrics_token(request: Request) -> str | None:
     authorization = request.headers.get("Authorization", "").strip()
     if authorization.lower().startswith("bearer "):
         return authorization[7:].strip() or None
-
     return None
 
 
@@ -121,574 +113,37 @@ async def _apply_status_rate_limit(request: Request) -> None:
     )
 
 
-def _to_int(value, default: int = 0) -> int:
-    try:
-        if value is None:
-            return default
-        return int(value)
-    except (TypeError, ValueError):
-        return default
+async def _collect_system_cache_metrics_payload(request: Request) -> dict:
+    from backend.presentation.routes import search as search_route
+    from backend.presentation.routes import tipi as tipi_route
 
+    nesh_internal = None
+    if hasattr(request.app.state, "service") and request.app.state.service:
+        nesh_internal = (
+            await request.app.state.service.snapshotNeshInternalCacheMetrics()
+        )
 
-def _normalize_db_status(raw_stats: dict | None, latency_ms: float) -> dict:
-    """Normaliza payload de status do banco principal para um contrato estável."""
-    if not raw_stats:
-        return {
-            "status": "error",
-            "chapters": 0,
-            "positions": 0,
-            "latency_ms": latency_ms,
-            "error": "Database unavailable",
-        }
+    tipi_internal = None
+    if hasattr(request.app.state, "tipi_service") and request.app.state.tipi_service:
+        tipi_internal = (
+            await request.app.state.tipi_service.snapshotTipiInternalCacheMetrics()
+        )
 
-    chapters = _to_int(raw_stats.get("chapters"))
-    positions = _to_int(raw_stats.get("positions"))
-    has_error = raw_stats.get("status") == "error"
-    payload = {
-        "status": "online"
-        if not has_error and chapters > 0 and positions > 0
-        else "error",
-        "chapters": chapters,
-        "positions": positions,
-        "latency_ms": latency_ms,
-    }
-    metadata = _extract_prefixed_metadata(raw_stats, "nesh")
-    if metadata:
-        payload["metadata"] = metadata
-    if raw_stats.get("error"):
-        payload["error"] = str(raw_stats.get("error"))
-    return payload
-
-
-def _normalize_tipi_status(raw_stats: dict | None) -> dict:
-    """Normaliza payload de status da TIPI para o mesmo contrato do banco principal."""
-    raw_stats = raw_stats or {}
-    chapters = _to_int(raw_stats.get("chapters"))
-    positions = _to_int(raw_stats.get("positions"))
-    is_online = bool(
-        (raw_stats.get("ok") is True or raw_stats.get("status") == "online")
-        and chapters > 0
-        and positions > 0
-    )
-
-    payload = {
-        "status": "online" if is_online else "error",
-        "chapters": chapters,
-        "positions": positions,
-    }
-    metadata = _extract_prefixed_metadata(raw_stats, "tipi")
-    if metadata:
-        payload["metadata"] = metadata
-    if raw_stats.get("error"):
-        payload["error"] = str(raw_stats.get("error"))
-    return payload
-
-
-def _extract_prefixed_metadata(raw_stats: dict | None, prefix: str) -> dict[str, str]:
-    metadata = (raw_stats or {}).get("metadata")
-    if not isinstance(metadata, dict):
-        return {}
-
-    prefix_token = f"{prefix}_"
     return {
-        key.removeprefix(prefix_token): str(value)
-        for key, value in metadata.items()
-        if key.startswith(prefix_token)
-    }
-
-
-def _normalize_count_catalog_status(
-    raw_stats: dict | None,
-    *,
-    count_field: str,
-    metadata_prefix: str,
-    public_count_field: str,
-) -> dict:
-    raw_stats = raw_stats or {}
-    total = _to_int(raw_stats.get(count_field))
-    payload = {
-        "status": (
-            "online" if raw_stats.get("status") != "error" and total > 0 else "error"
+        "status": "ok",
+        "search_code_payload_cache": (
+            search_route.snapshotSearchCodePayloadCacheMetrics()
         ),
-        public_count_field: total,
+        "tipi_code_payload_cache": (
+            tipi_route.snapshotTipiCodePayloadCacheMetrics()
+        ),
+        "nesh_internal_caches": nesh_internal,
+        "tipi_internal_caches": tipi_internal,
     }
-    metadata = _extract_prefixed_metadata(raw_stats, metadata_prefix)
-    if metadata:
-        payload["metadata"] = metadata
-    if raw_stats.get("error"):
-        payload["error"] = str(raw_stats.get("error"))
-    return payload
-
-
-async def _collect_db_status(request: Request) -> tuple[dict, float]:
-    db = getattr(request.app.state, "db", None)
-    start = time.perf_counter()
-
-    if db:
-        db_stats = await db.check_connection()
-        latency_ms = round((time.perf_counter() - start) * 1000, 2)
-        return db_stats, latency_ms
-
-    try:
-        from sqlalchemy import text
-
-        from backend.infrastructure.db_engine import get_session
-
-        async with get_session() as session:
-            chapters_count = await session.execute(
-                text("SELECT COUNT(*) FROM chapters")
-            )
-            positions_count = await session.execute(
-                text("SELECT COUNT(*) FROM positions")
-            )
-            metadata: dict[str, str] = {}
-            if settings.database.is_postgres:
-                try:
-                    metadata_result = await session.execute(
-                        text(
-                            """
-                            SELECT key, value
-                            FROM catalog_metadata
-                            WHERE key LIKE 'nesh_%'
-                            ORDER BY key
-                            """
-                        )
-                    )
-                    metadata = {row.key: row.value for row in metadata_result}
-                except Exception:
-                    metadata = {}
-        db_stats = {
-            "status": "online",
-            "chapters": int(chapters_count.scalar() or 0),
-            "positions": int(positions_count.scalar() or 0),
-            "metadata": metadata,
-        }
-    except Exception as e:
-        db_stats = {"status": "error", "error": str(e)}
-
-    latency_ms = round((time.perf_counter() - start) * 1000, 2)
-    return db_stats, latency_ms
-
-
-async def _collect_tipi_status(request: Request) -> dict:
-    tipi_service = getattr(request.app.state, "tipi_service", None)
-    if tipi_service is None:
-        return {"status": "error", "error": "TIPI service unavailable"}
-
-    try:
-        return await tipi_service.check_connection()
-    except Exception as tipi_err:
-        return {"status": "error", "error": str(tipi_err)}
-
-
-async def _collect_nbs_status(request: Request) -> dict:
-    nbs_service = getattr(request.app.state, "nbs_service", None)
-    if nbs_service is None:
-        return {"status": "error", "error": "NBS service unavailable"}
-
-    try:
-        return await nbs_service.check_connection()
-    except Exception as nbs_err:
-        return {"status": "error", "error": str(nbs_err)}
-
-
-async def _collect_status_payloads_uncached(
-    request: Request,
-) -> tuple[dict, dict, dict, dict, str]:
-    db_stats, db_latency_ms = await _collect_db_status(request)
-    tipi_stats = await _collect_tipi_status(request)
-    nbs_stats = await _collect_nbs_status(request)
-
-    normalized_db = _normalize_db_status(db_stats, db_latency_ms)
-    normalized_tipi = _normalize_tipi_status(tipi_stats)
-    normalized_nbs = _normalize_count_catalog_status(
-        nbs_stats,
-        count_field="nbs_items",
-        metadata_prefix="nbs",
-        public_count_field="items",
-    )
-    normalized_nebs = _normalize_count_catalog_status(
-        nbs_stats,
-        count_field="nebs_entries",
-        metadata_prefix="nebs",
-        public_count_field="entries",
-    )
-    overall_status = (
-        "online"
-        if normalized_db.get("status") == "online"
-        and normalized_tipi.get("status") == "online"
-        and normalized_nbs.get("status") == "online"
-        and normalized_nebs.get("status") == "online"
-        else "error"
-    )
-    return (
-        normalized_db,
-        normalized_tipi,
-        normalized_nbs,
-        normalized_nebs,
-        overall_status,
-    )
-
-
-def _build_status_snapshot(
-    normalized_db: dict,
-    normalized_tipi: dict,
-    normalized_nbs: dict,
-    normalized_nebs: dict,
-    overall_status: str,
-) -> dict:
-    return {
-        "normalized_db": normalized_db,
-        "normalized_tipi": normalized_tipi,
-        "normalized_nbs": normalized_nbs,
-        "normalized_nebs": normalized_nebs,
-        "overall_status": overall_status,
-    }
-
-
-def _unpack_status_snapshot(snapshot: dict) -> tuple[dict, dict, dict, dict, str]:
-    return (
-        dict(snapshot.get("normalized_db") or {}),
-        dict(snapshot.get("normalized_tipi") or {}),
-        dict(snapshot.get("normalized_nbs") or {}),
-        dict(snapshot.get("normalized_nebs") or {}),
-        str(snapshot.get("overall_status") or "error"),
-    )
-
-
-def _store_status_snapshot(snapshot: dict, *, expires_at: float) -> dict:
-    _STATUS_CACHE["value"] = snapshot
-    _STATUS_CACHE["expires_at"] = expires_at
-    return snapshot
-
-
-async def _refresh_status_snapshot(request: Request, ttl_seconds: int) -> dict:
-    snapshot = _build_status_snapshot(*await _collect_status_payloads_uncached(request))
-    _store_status_snapshot(snapshot, expires_at=time.monotonic() + ttl_seconds)
-    if redis_cache.available:
-        await redis_cache.set_status_snapshot("public", snapshot)
-    return snapshot
-
-
-async def _read_redis_status_snapshot(*, now: float, ttl_seconds: int) -> dict | None:
-    if not redis_cache.available:
-        return None
-
-    redis_cached = await redis_cache.get_status_snapshot("public")
-    if not isinstance(redis_cached, dict):
-        return None
-
-    return _store_status_snapshot(redis_cached, expires_at=now + ttl_seconds)
-
-
-async def _await_status_refresh_snapshot(request: Request, ttl_seconds: int) -> dict:
-    global _STATUS_CACHE_REFRESH_TASK
-
-    task: asyncio.Task | None = None
-    lock = _get_status_cache_lock()
-    async with lock:
-        cached = _read_l1_status_snapshot()
-        if cached is not None:
-            return cached
-
-        task = _STATUS_CACHE_REFRESH_TASK
-        if task is None:
-            task = asyncio.create_task(_refresh_status_snapshot(request, ttl_seconds))
-            _STATUS_CACHE_REFRESH_TASK = task
-
-    try:
-        return await task
-    finally:
-        async with lock:
-            if _STATUS_CACHE_REFRESH_TASK is task:
-                _STATUS_CACHE_REFRESH_TASK = None
-
-
-async def _recover_status_snapshot(ttl_seconds: int) -> dict | None:
-    stale = _read_stale_l1_status_snapshot()
-    if stale is not None:
-        return stale
-
-    return await _read_redis_status_snapshot(
-        now=time.monotonic(),
-        ttl_seconds=ttl_seconds,
-    )
-
-
-async def _get_status_snapshot(request: Request) -> dict:
-    ttl_seconds = _status_cache_ttl_seconds()
-    if ttl_seconds <= 0:
-        return _build_status_snapshot(*await _collect_status_payloads_uncached(request))
-
-    now = time.monotonic()
-    cached = _read_l1_status_snapshot(now)
-    if cached is not None:
-        return cached
-
-    redis_cached = await _read_redis_status_snapshot(now=now, ttl_seconds=ttl_seconds)
-    if redis_cached is not None:
-        return redis_cached
-
-    try:
-        return await _await_status_refresh_snapshot(request, ttl_seconds)
-    except Exception:
-        fallback = await _recover_status_snapshot(ttl_seconds)
-        if fallback is not None:
-            return fallback
-        raise
-
-
-async def _collect_status_payloads(
-    request: Request,
-) -> tuple[dict, dict, dict, dict, str]:
-    return _unpack_status_snapshot(await _get_status_snapshot(request))
-
-
-def _build_public_status_payload(
-    normalized_db: dict,
-    normalized_tipi: dict,
-    normalized_nbs: dict | str | None = None,
-    normalized_nebs: dict | None = None,
-    overall_status: str | None = None,
-) -> dict:
-    legacy_mode = False
-    if isinstance(normalized_nbs, str) and overall_status is None:
-        overall_status = normalized_nbs
-        normalized_nbs = None
-        normalized_nebs = None
-        legacy_mode = True
-
-    overall_status = overall_status or "error"
-    catalogs = {
-        "nesh": {"status": normalized_db.get("status", "error")},
-        "tipi": {"status": normalized_tipi.get("status", "error")},
-    }
-    if normalized_nbs is not None:
-        catalogs["nbs"] = {"status": normalized_nbs.get("status", "error")}
-    if normalized_nebs is not None:
-        catalogs["nebs"] = {"status": normalized_nebs.get("status", "error")}
-    payload = {
-        "status": overall_status,
-        "database": {"status": normalized_db.get("status", "error")},
-        "tipi": {
-            "status": normalized_tipi.get("status", "error"),
-        },
-    }
-    if not legacy_mode:
-        payload["catalogs"] = catalogs
-    if normalized_nbs is not None:
-        payload["nbs"] = {"status": normalized_nbs.get("status", "error")}
-    if normalized_nebs is not None:
-        payload["nebs"] = {"status": normalized_nebs.get("status", "error")}
-    return payload
-
-
-def _build_detailed_status_payload(
-    request: Request,
-    normalized_db: dict,
-    normalized_tipi: dict,
-    normalized_nbs: dict,
-    normalized_nebs: dict,
-    overall_status: str,
-) -> dict:
-    catalogs = {
-        "nesh": normalized_db,
-        "tipi": normalized_tipi,
-        "nbs": normalized_nbs,
-        "nebs": normalized_nebs,
-    }
-    return {
-        "status": overall_status,
-        "version": getattr(request.app, "version", None),
-        "backend": "FastAPI",
-        "database": normalized_db,
-        "tipi": normalized_tipi,
-        "nbs": {
-            "status": normalized_nbs.get("status", "error"),
-            "items": int(normalized_nbs.get("items") or 0),
-            **(
-                {"metadata": normalized_nbs["metadata"]}
-                if isinstance(normalized_nbs.get("metadata"), dict)
-                else {}
-            ),
-            **(
-                {"error": normalized_nbs["error"]}
-                if normalized_nbs.get("error")
-                else {}
-            ),
-        },
-        "nebs": {
-            "status": normalized_nebs.get("status", "error"),
-            "entries": int(normalized_nebs.get("entries") or 0),
-            **(
-                {"metadata": normalized_nebs["metadata"]}
-                if isinstance(normalized_nebs.get("metadata"), dict)
-                else {}
-            ),
-            **(
-                {"error": normalized_nebs["error"]}
-                if normalized_nebs.get("error")
-                else {}
-            ),
-        },
-        "catalogs": catalogs,
-    }
-
-
-def _metric_value_from_status(status: str) -> int:
-    return 1 if status == "online" else 0
-
-
-def _append_metric_line(
-    lines: list[str],
-    name: str,
-    value: int | float,
-    labels: dict[str, str] | None = None,
-) -> None:
-    label_text = ""
-    if labels:
-
-        def _escape_label_value(label_value: str) -> str:
-            return (
-                label_value.replace("\\", "\\\\")
-                .replace('"', '\\"')
-                .replace("\n", "\\n")
-            )
-
-        encoded_labels = ",".join(
-            f'{key}="{_escape_label_value(str(label_value))}"'
-            for key, label_value in sorted(labels.items())
-        )
-        label_text = f"{{{encoded_labels}}}"
-    lines.append(f"{name}{label_text} {value}")
-
-
-def _append_catalog_status_metrics(lines: list[str], status_payload: dict) -> None:
-    catalogs = status_payload.get("catalogs", {})
-    for catalog_name, catalog_payload in catalogs.items():
-        if not isinstance(catalog_payload, dict):
-            continue
-        _append_metric_line(
-            lines,
-            "nesh_catalog_status",
-            _metric_value_from_status(str(catalog_payload.get("status", "error"))),
-            {"catalog": catalog_name},
-        )
-
-
-def _append_database_latency_metric(lines: list[str], status_payload: dict) -> None:
-    database_payload = status_payload.get("database", {})
-    if not isinstance(database_payload, dict) or "latency_ms" not in database_payload:
-        return
-
-    lines.extend(
-        [
-            "# HELP nesh_database_latency_ms Database latency observed by /api/status/details.",
-            "# TYPE nesh_database_latency_ms gauge",
-        ]
-    )
-    _append_metric_line(
-        lines,
-        "nesh_database_latency_ms",
-        float(database_payload.get("latency_ms") or 0),
-    )
-
-
-def _append_payload_cache_metrics(
-    lines: list[str],
-    cache_metrics: dict,
-    *,
-    metric_name: str,
-    help_text: str,
-    field_name: str,
-) -> None:
-    lines.extend(
-        [
-            f"# HELP {metric_name} {help_text}",
-            f"# TYPE {metric_name} gauge",
-        ]
-    )
-    for cache_name in ("search_code_payload_cache", "tipi_code_payload_cache"):
-        cache_payload = cache_metrics.get(cache_name)
-        if not isinstance(cache_payload, dict):
-            continue
-        _append_metric_line(
-            lines,
-            metric_name,
-            int(cache_payload.get(field_name) or 0),
-            {"cache": cache_name},
-        )
-
-
-def _append_internal_cache_hit_rate_metrics(
-    lines: list[str], cache_metrics: dict
-) -> None:
-    lines.extend(
-        [
-            "# HELP nesh_internal_cache_hit_rate Internal service cache hit rate.",
-            "# TYPE nesh_internal_cache_hit_rate gauge",
-        ]
-    )
-    for service_name in ("nesh_internal_caches", "tipi_internal_caches"):
-        service_payload = cache_metrics.get(service_name)
-        if not isinstance(service_payload, dict):
-            continue
-        for cache_name, cache_payload in service_payload.items():
-            if not isinstance(cache_payload, dict):
-                continue
-            hit_rate = cache_payload.get("hit_rate")
-            if not isinstance(hit_rate, (int, float)):
-                continue
-            _append_metric_line(
-                lines,
-                "nesh_internal_cache_hit_rate",
-                float(hit_rate),
-                {"service": service_name, "cache": cache_name},
-            )
-
-
-def _build_prometheus_metrics_payload(
-    status_payload: dict,
-    cache_metrics: dict,
-) -> str:
-    lines: list[str] = [
-        "# HELP nesh_catalog_status Catalog health status (1=online, 0=error).",
-        "# TYPE nesh_catalog_status gauge",
-    ]
-    _append_catalog_status_metrics(lines, status_payload)
-    _append_database_latency_metric(lines, status_payload)
-    _append_payload_cache_metrics(
-        lines,
-        cache_metrics,
-        metric_name="nesh_payload_cache_hits",
-        help_text="Total payload-cache hits by route family.",
-        field_name="hits",
-    )
-    _append_payload_cache_metrics(
-        lines,
-        cache_metrics,
-        metric_name="nesh_payload_cache_misses",
-        help_text="Total payload-cache misses by route family.",
-        field_name="misses",
-    )
-    _append_internal_cache_hit_rate_metrics(lines, cache_metrics)
-
-    return "\n".join(lines) + "\n"
 
 
 @router.get("/status", responses=STATUS_RESPONSES)
-async def get_status(request: Request):
-    """
-    Healthcheck e Status do Sistema.
-
-    Verifica conectividade com:
-    - Catálogo principal NESH
-    - TIPI
-    - NBS
-    - NEBS
-
-    Retorna apenas o mínimo necessário para readiness público.
-    """
+async def fetch_system_status(request: Request):
     await _apply_status_rate_limit(request)
     (
         normalized_db,
@@ -696,8 +151,8 @@ async def get_status(request: Request):
         normalized_nbs,
         normalized_nebs,
         overall_status,
-    ) = await _collect_status_payloads(request)
-    return _build_public_status_payload(
+    ) = await collect_status_payloads(request)
+    return build_public_status_payload(
         normalized_db,
         normalized_tipi,
         normalized_nbs,
@@ -707,16 +162,13 @@ async def get_status(request: Request):
 
 
 @router.head("/status", include_in_schema=False)
-async def head_status(request: Request):
-    await _collect_status_payloads(request)
+async def head_system_status(request: Request):
+    await collect_status_payloads(request)
     return Response(status_code=200)
 
 
 @router.get("/status/details", responses=STATUS_DETAILS_RESPONSES)
-async def get_status_details(request: Request):
-    """
-    Status detalhado do sistema para administradores.
-    """
+async def fetch_system_status_details(request: Request):
     await _apply_status_rate_limit(request)
     if not await _is_admin_request(request):
         raise HTTPException(status_code=403, detail="Forbidden")
@@ -727,8 +179,8 @@ async def get_status_details(request: Request):
         normalized_nbs,
         normalized_nebs,
         overall_status,
-    ) = await _collect_status_payloads(request)
-    return _build_detailed_status_payload(
+    ) = await collect_status_payloads(request)
+    return build_detailed_status_payload(
         request,
         normalized_db,
         normalized_tipi,
@@ -738,42 +190,14 @@ async def get_status_details(request: Request):
     )
 
 
-async def _collect_cache_metrics_payload(request: Request) -> dict:
-    from backend.presentation.routes import search as search_route
-    from backend.presentation.routes import tipi as tipi_route
-
-    nesh_internal = None
-    if hasattr(request.app.state, "service") and request.app.state.service:
-        nesh_internal = await request.app.state.service.get_internal_cache_metrics()
-
-    tipi_internal = None
-    if hasattr(request.app.state, "tipi_service") and request.app.state.tipi_service:
-        tipi_internal = (
-            await request.app.state.tipi_service.get_internal_cache_metrics()
-        )
-
-    return {
-        "status": "ok",
-        "search_code_payload_cache": search_route.get_payload_cache_metrics(),
-        "tipi_code_payload_cache": tipi_route.get_payload_cache_metrics(),
-        "nesh_internal_caches": nesh_internal,
-        "tipi_internal_caches": tipi_internal,
-    }
-
-
 @router.get(
     "/cache-metrics",
     responses={403: {"description": "Forbidden (admin-only endpoint)."}},
 )
-async def get_cache_metrics(request: Request):
-    """
-    Métricas de hit/miss dos payload caches de /api/search e /api/tipi/search.
-    Restrito a admins por conter dados operacionais internos.
-    """
+async def fetch_system_cache_metrics(request: Request):
     if not await _is_admin_request(request):
         raise HTTPException(status_code=403, detail="Forbidden")
-
-    return await _collect_cache_metrics_payload(request)
+    return await _collect_system_cache_metrics_payload(request)
 
 
 @router.get(
@@ -782,11 +206,7 @@ async def get_cache_metrics(request: Request):
     response_class=PlainTextResponse,
     responses=METRICS_RESPONSES,
 )
-async def get_metrics(request: Request):
-    """
-    Endpoint Prometheus-style para observabilidade básica.
-    Protegido por token dedicado em X-Metrics-Token ou Authorization: Bearer.
-    """
+async def fetch_system_metrics(request: Request):
     if not _metrics_endpoint_enabled():
         raise HTTPException(status_code=404, detail="Not Found")
     if not _is_metrics_request_authorized(request):
@@ -799,8 +219,8 @@ async def get_metrics(request: Request):
         normalized_nbs,
         normalized_nebs,
         overall_status,
-    ) = await _collect_status_payloads(request)
-    status_payload = _build_detailed_status_payload(
+    ) = await collect_status_payloads(request)
+    status_payload = build_detailed_status_payload(
         request,
         normalized_db,
         normalized_tipi,
@@ -808,8 +228,8 @@ async def get_metrics(request: Request):
         normalized_nebs,
         overall_status,
     )
-    cache_metrics = await _collect_cache_metrics_payload(request)
-    payload = _build_prometheus_metrics_payload(status_payload, cache_metrics)
+    cache_metrics = await _collect_system_cache_metrics_payload(request)
+    payload = build_prometheus_metrics_payload(status_payload, cache_metrics)
     return PlainTextResponse(
         payload,
         media_type="text/plain; version=0.0.4; charset=utf-8",
@@ -823,30 +243,21 @@ async def get_metrics(request: Request):
         404: {"description": "Not found when debug mode is disabled."},
     },
 )
-async def debug_anchors(
+async def debug_nesh_anchors(
     request: Request,
     service: Annotated[NeshService, Depends(get_nesh_service)],
     ncm: Annotated[str, Query(description="Código NCM para debug de anchors")],
 ):
-    """
-    DEBUG: Retorna o HTML renderizado e lista todos os IDs injetados.
-    Útil para diagnosticar problemas de scroll.
-    """
     if not settings.features.debug_mode:
         raise HTTPException(status_code=404, detail="Not found")
-
     if not await _is_admin_request(request):
         raise HTTPException(status_code=403, detail="Forbidden")
 
-    response_data = await service.process_request(ncm)
-
-    # Collect all IDs from the rendered HTML
+    response_data = await service.executeNeshSearchWithVectorWeights(ncm)
     html_content = response_data.get("markdown", "") or ""
     id_pattern = re.compile(r'id="([^"]+)"')
     all_ids = id_pattern.findall(html_content)
-
-    # Filter to position-related IDs
-    pos_ids = [id for id in all_ids if id.startswith("pos-") or id.startswith("cap-")]
+    pos_ids = [item for item in all_ids if item.startswith("pos-") or item.startswith("cap-")]
 
     return {
         "query": ncm,
@@ -863,12 +274,8 @@ async def debug_anchors(
     "/admin/reload-secrets",
     responses={403: {"description": "Forbidden (admin-only endpoint)."}},
 )
-async def reload_secrets(request: Request):
-    """
-    Recarrega secrets de env/.env sem reiniciar o servidor.
-    """
+async def reload_system_secrets(request: Request):
     if not await _is_admin_request(request):
         raise HTTPException(status_code=403, detail="Forbidden")
-
     reload_settings()
     return {"success": True}
