+++ backend/server/app.py
@@ -1,9 +1,14 @@
-import json
+from __future__ import annotations
+
 import logging
-import os
+from collections.abc import Awaitable, Callable
 from contextlib import asynccontextmanager
-from typing import Any, cast
-from urllib.parse import urlparse
+from typing import Any, ParamSpec, TypeVar, cast
+
+from fastapi import FastAPI, Request, Response
+from fastapi.middleware.cors import CORSMiddleware
+from fastapi.middleware.gzip import GZipMiddleware
+from fastapi.responses import JSONResponse
 
 from backend.config import CONFIG, setup_logging
 from backend.config.exceptions import NeshError
@@ -11,18 +16,63 @@ from backend.config.settings import settings
 from backend.data.glossary_manager import init_glossary
 from backend.infrastructure import DatabaseAdapter
 from backend.infrastructure.redis_client import redis_cache
-
-# Import New Routers
-from backend.presentation.routes import (
-    auth,
-    comments,
-    database_download,
-    profile,
-    search,
-    services,
-    system,
-    tipi,
-    webhooks,
+from backend.server.app_lifecycle import (
+    _collect_release_metadata as _collect_release_metadata_impl,
+)
+from backend.server.app_lifecycle import (
+    _first_non_empty_env as _first_non_empty_env_impl,
+)
+from backend.server.app_lifecycle import (
+    _init_cache_warmup as _init_cache_warmup_impl,
+)
+from backend.server.app_lifecycle import (
+    _init_nbs_service as _init_nbs_service_impl,
+)
+from backend.server.app_lifecycle import (
+    _init_nesh_service as _init_nesh_service_impl,
+)
+from backend.server.app_lifecycle import (
+    _init_primary_database as _init_primary_database_impl,
+)
+from backend.server.app_lifecycle import (
+    _init_sqlmodel_engine as _init_sqlmodel_engine_impl,
+)
+from backend.server.app_lifecycle import (
+    _init_tipi_service as _init_tipi_service_impl,
+)
+from backend.server.app_lifecycle import (
+    _log_runtime_security_warnings as _log_runtime_security_warnings_impl,
+)
+from backend.server.app_lifecycle import (
+    _postgres_tipi_has_data as _postgres_tipi_has_data_impl,
+)
+from backend.server.app_lifecycle import (
+    _record_release_metadata as _record_release_metadata_impl,
+)
+from backend.server.app_lifecycle import (
+    _resolve_project_root as _resolve_project_root_impl,
+)
+from backend.server.app_lifecycle import (
+    _shutdown_resources as _shutdown_resources_impl,
+)
+from backend.server.app_lifecycle import (
+    _validate_dev_tenant_override_safety as _validate_dev_tenant_override_safety_impl,
+)
+from backend.server.app_routes import _configure_routes
+from backend.server.app_security import (
+    _apply_security_headers as _apply_security_headers_impl,
+)
+from backend.server.app_security import (
+    _build_content_security_policy as _build_content_security_policy_impl,
+)
+from backend.server.app_security import (
+    _build_cors_configuration as _build_cors_configuration_impl,
+)
+from backend.server.app_security import (
+    _request_uses_https as _request_uses_https_impl,
+)
+from backend.server.app_security import (
+    _should_expose_api_docs as _should_expose_api_docs_impl,
 )
 from backend.server.error_handlers import (
     generic_exception_handler,
@@ -33,381 +83,114 @@ from backend.server.middleware import (
     is_loopback_host,
     origin_looks_like_loopback,
 )
-from backend.server.observability import configure_observability
 from backend.services.ai_service import AiService
 from backend.services.nbs_service import NbsService
 from backend.services.tipi_service import TipiService
+from backend.server.observability import configure_observability
 from backend.utils.frontend_check import verify_frontend_build
-from fastapi import FastAPI, Request, Response
-from fastapi.middleware.cors import CORSMiddleware
-from fastapi.middleware.gzip import GZipMiddleware
-from fastapi.responses import JSONResponse
-from fastapi.staticfiles import StaticFiles
 
-"""
-Módulo do Servidor (API Handler).
+setup_logging()
+logger = logging.getLogger("server")
 
-Define a aplicação FastAPI, rotas da API e ciclo de vida do servidor.
-Responsável por:
-1. Inicializar recursos globais (DB, Services) no startup.
-2. Definir endpoints REST para busca NCM e TIPI (via Routers).
-3. Servir o frontend React compilado (arquivos estáticos).
-4. Gerenciar tratamento de erros e respostas JSON.
-"""
-
-# Logger setup
-setup_logging(
-    level=settings.logging.python_level,
-    redact_sensitive_data=settings.logging.redact_sensitive_data,
-)
-logger = logging.getLogger("nesh.server")
-
-_DOCS_PATHS = frozenset({"/docs", "/redoc", "/openapi.json"})
+P = ParamSpec("P")
+R = TypeVar("R")
 
 
-def _build_content_security_policy() -> str:
-    connect_sources = ["'self'", "https:", "wss:"]
-    if settings.server.env == "development":
-        connect_sources.extend(
-            [
-                "http://127.0.0.1:8000",
-                "http://localhost:8000",
-                "ws://127.0.0.1:*",
-                "ws://localhost:*",
-            ]
-        )
+def _sync_lifecycle_dependencies() -> None:
+    """Keeps the lifecycle module aligned with monkeypatched app-level symbols."""
+    import backend.server.app_lifecycle as app_lifecycle
 
-    return "; ".join(
-        (
-            "default-src 'self'",
-            "base-uri 'self'",
-            "object-src 'none'",
-            "frame-ancestors 'none'",
-            "form-action 'self'",
-            (
-                "script-src 'self' https://*.clerk.accounts.dev "
-                "https://*.clerk.com https://challenges.cloudflare.com"
-            ),
-            "style-src 'self' 'unsafe-inline' https://fonts.googleapis.com",
-            "img-src 'self' data: blob: https:",
-            "font-src 'self' https://fonts.gstatic.com data:",
-            f"connect-src {' '.join(connect_sources)}",
-            "worker-src 'self' blob:",
-            (
-                "frame-src 'self' https://*.clerk.accounts.dev "
-                "https://*.clerk.com https://challenges.cloudflare.com"
-            ),
-        )
-    )
+    app_lifecycle.CONFIG = CONFIG
+    app_lifecycle.settings = settings
+    app_lifecycle.DatabaseAdapter = DatabaseAdapter
+    app_lifecycle.redis_cache = redis_cache
+    app_lifecycle.init_glossary = init_glossary
+    app_lifecycle.verify_frontend_build = verify_frontend_build
+    app_lifecycle.AiService = AiService
+    app_lifecycle.NbsService = NbsService
+    app_lifecycle.TipiService = TipiService
+    app_lifecycle.configure_observability = configure_observability
+    app_lifecycle.is_loopback_host = is_loopback_host
+    app_lifecycle.origin_looks_like_loopback = origin_looks_like_loopback
+    app_lifecycle.logger = logger
+
+
+def _sync_lifecycle_call(callback: Callable[P, R]) -> Callable[P, R]:
+    def wrapped(*args: P.args, **kwargs: P.kwargs) -> R:
+        _sync_lifecycle_dependencies()
+        return callback(*args, **kwargs)
+
+    return wrapped
+
+
+def _sync_lifecycle_async_call(
+    callback: Callable[P, Awaitable[R]],
+) -> Callable[P, Awaitable[R]]:
+    async def wrapped(*args: P.args, **kwargs: P.kwargs) -> R:
+        _sync_lifecycle_dependencies()
+        return await callback(*args, **kwargs)
+
+    return wrapped
 
 
 def _resolve_project_root() -> str:
-    return os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
-
-def _validate_dev_tenant_override_safety() -> None:
-    if settings.server.env != "development" or not settings.features.debug_mode:
-        return
-
-    if is_loopback_host(settings.server.host):
-        return
-
-    raise RuntimeError(
-        "Debug tenant overrides require a localhost-only host binding. "
-        f"Received SERVER__HOST={settings.server.host!r}."
-    )
-
-
-def _should_expose_api_docs() -> bool:
-    return settings.server.env == "development" and settings.features.debug_mode
+    return _sync_lifecycle_call(_resolve_project_root_impl)(__file__)
 
 
 def _first_non_empty_env(*names: str) -> str | None:
-    for name in names:
-        value = os.getenv(name, "").strip()
-        if value:
-            return value
-    return None
+    return _first_non_empty_env_impl(*names)
 
 
-def _collect_release_metadata(app: FastAPI) -> dict[str, str]:
-    metadata: dict[str, str] = {
-        "app_version": getattr(app, "version", "unknown"),
-        "server_env": settings.server.env,
-    }
+_collect_release_metadata = _sync_lifecycle_call(_collect_release_metadata_impl)
+_record_release_metadata = _sync_lifecycle_call(_record_release_metadata_impl)
 
-    direct_env_map = {
-        "render_service_id": "RENDER_SERVICE_ID",
-        "render_service_name": "RENDER_SERVICE_NAME",
-        "render_instance_id": "RENDER_INSTANCE_ID",
-        "render_external_url": "RENDER_EXTERNAL_URL",
-        "render_region": "RENDER_REGION",
-    }
-    for field, env_name in direct_env_map.items():
-        value = os.getenv(env_name, "").strip()
-        if value:
-            metadata[field] = value
 
-    git_commit = _first_non_empty_env(
-        "RENDER_GIT_COMMIT",
-        "SOURCE_VERSION",
-        "GIT_COMMIT",
-        "COMMIT_SHA",
+def _build_content_security_policy() -> str:
+    return _build_content_security_policy_impl(settings.server.env)
+
+
+def _should_expose_api_docs() -> bool:
+    return _should_expose_api_docs_impl(
+        settings.server.env,
+        settings.features.debug_mode,
     )
-    if git_commit:
-        metadata["git_commit"] = git_commit
-
-    git_branch = _first_non_empty_env(
-        "RENDER_GIT_BRANCH",
-        "GIT_BRANCH",
-        "BRANCH",
-    )
-    if git_branch:
-        metadata["git_branch"] = git_branch
-
-    deploy_id = _first_non_empty_env(
-        "RENDER_DEPLOY_ID",
-        "RENDER_BUILD_ID",
-        "DEPLOY_ID",
-    )
-    if deploy_id:
-        metadata["deploy_id"] = deploy_id
-
-    return metadata
-
-
-def _record_release_metadata(app: FastAPI) -> None:
-    metadata = _collect_release_metadata(app)
-    app.state.release_metadata = metadata
-    logger.info(
-        "Runtime release metadata: %s",
-        json.dumps(metadata, ensure_ascii=False, sort_keys=True),
-    )
-    configure_observability(
-        release=metadata.get("git_commit") or metadata.get("app_version"),
-        server_name=metadata.get("render_service_name")
-        or metadata.get("render_service_id"),
-    )
-
-
-def _log_runtime_security_warnings() -> None:
-    if settings.server.env != "production":
-        return
-
-    warnings: list[str] = []
-    if settings.features.debug_mode:
-        warnings.append("SERVER__ENV=production com FEATURES__DEBUG_MODE=true.")
-
-    if not settings.server.cors_allowed_origins:
-        warnings.append(
-            "SERVER__CORS_ALLOWED_ORIGINS vazio em produção; configure apenas domínios oficiais."
-        )
-    elif any(origin.strip() == "*" for origin in settings.server.cors_allowed_origins):
-        warnings.append(
-            'SERVER__CORS_ALLOWED_ORIGINS="*" é inseguro em produção; configure apenas origens explícitas.'
-        )
-    elif any(
-        origin_looks_like_loopback(origin)
-        for origin in settings.server.cors_allowed_origins
-    ):
-        warnings.append(
-            "SERVER__CORS_ALLOWED_ORIGINS em produção ainda inclui origem localhost/loopback."
-        )
-
-    redis_hostname = urlparse(settings.cache.redis_url).hostname
-    if settings.cache.enable_redis and is_loopback_host(redis_hostname):
-        warnings.append(
-            "CACHE__ENABLE_REDIS=true em produção com CACHE__REDIS_URL apontando para localhost."
-        )
-
-    if not settings.database.is_postgres:
-        warnings.append(
-            "DATABASE__ENGINE não está em postgresql no ambiente de produção."
-        )
-
-    for warning in warnings:
-        logger.warning("Runtime security warning: %s", warning)
 
 
 def _request_uses_https(request: Request) -> bool:
-    forwarded_proto = request.headers.get("x-forwarded-proto", "")
-    if forwarded_proto:
-        proto = forwarded_proto.split(",", 1)[0].strip().lower()
-        if proto == "https":
-            return True
-    return request.url.scheme.lower() == "https"
+    return _request_uses_https_impl(request)
 
 
-def _apply_security_headers(request: Request, response) -> None:
-    response.headers["Content-Security-Policy"] = _build_content_security_policy()
-    response.headers["X-Frame-Options"] = "DENY"
-    response.headers["X-Content-Type-Options"] = "nosniff"
-    response.headers["Referrer-Policy"] = "strict-origin-when-cross-origin"
-    response.headers["Permissions-Policy"] = "camera=(), microphone=(), geolocation=()"
-    if _request_uses_https(request):
-        response.headers["Strict-Transport-Security"] = (
-            "max-age=63072000; includeSubDomains; preload"
-        )
+def _apply_security_headers(request: Request, response: Response) -> None:
+    _apply_security_headers_impl(request, response, settings.server.env)
 
 
-async def _init_primary_database(app: FastAPI) -> None:
-    logger.info("Initializing Database...")
-    if settings.database.is_postgres:
-        # Postgres path usa SQLAlchemy/Repository; o adapter legado não é necessário.
-        app.state.db = None
-        return
-
-    app.state.db = DatabaseAdapter(CONFIG.db_path)
-    # Cria pool cedo para falhar rápido em startup caso o arquivo/caminho esteja inválido.
-    await app.state.db._ensure_pool()
-
-
-async def _init_sqlmodel_engine(app: FastAPI) -> None:
-    if settings.database.is_postgres:
-        # Em Postgres, migrations são responsabilidade do Alembic.
-        app.state.sqlmodel_enabled = True
-        logger.info("SQLModel engine ready (Postgres migrations via Alembic)")
-        return
-
-    # SQLModel é opcional no modo legado; tratamos fallback para não bloquear startup em SQLite.
-    try:
-        from backend.infrastructure.db_engine import init_db
-    except ImportError:
-        app.state.sqlmodel_enabled = False
-        logger.debug("SQLModel not available, using legacy DatabaseAdapter")
-        return
-
-    try:
-        await init_db()
-        app.state.sqlmodel_enabled = True
-        logger.info("SQLModel engine initialized (SQLite)")
-    except Exception as e:
-        app.state.sqlmodel_enabled = False
-        logger.warning("SQLModel init skipped (SQLite incompatibility): %s", e)
-
-
-async def _init_nesh_service(app: FastAPI) -> None:
-    logger.info("Initializing Services...")
-    from backend.services.nesh_service import NeshService
-
-    if settings.database.is_postgres:
-        # Repository mode suporta RLS e isolamento por tenant no Postgres.
-        app.state.service = await NeshService.create_with_repository()
-        logger.info("NeshService initialized in Repository mode (Postgres/RLS)")
-        return
-
-    if app.state.db is None:
-        raise RuntimeError("DatabaseAdapter não inicializado para modo SQLite")
-    app.state.service = NeshService(app.state.db)
-    logger.info("NeshService initialized in Legacy mode (SQLite)")
-
-
-async def _init_cache_warmup(app: FastAPI) -> None:
-    if not settings.cache.enable_redis:
-        return
-
-    try:
-        await redis_cache.connect()
-    except Exception as e:
-        logger.warning("Redis connect failed during startup: %s", e)
-        return
-
-    if not redis_cache.available:
-        return
-
-    try:
-        warmed = await app.state.service.prewarm_cache()
-        logger.info("Chapter cache prewarmed: %s capítulos", warmed)
-    except Exception as e:
-        # Prewarm é otimização: não deve derrubar startup.
-        logger.warning("Cache prewarm failed: %s", e)
-
-
-async def _postgres_tipi_has_data() -> bool:
-    try:
-        from backend.infrastructure.db_engine import get_session
-        from sqlalchemy import text
-
-        async with get_session() as session:
-            result = await session.execute(
-                text("SELECT EXISTS(SELECT 1 FROM tipi_positions)")
-            )
-            has_data = bool(result.scalar())
-        logger.info("TIPI PostgreSQL data available: %s", has_data)
-        return has_data
-    except Exception as e:
-        logger.warning("Could not check tipi_positions table: %s", e)
-        return False
-
-
-async def _init_tipi_service(app: FastAPI) -> None:
-    if not settings.database.is_postgres:
-        app.state.tipi_service = TipiService()
-        logger.info("TipiService initialized in Legacy mode (SQLite)")
-        return
-
-    if await _postgres_tipi_has_data():
-        app.state.tipi_service = await TipiService.create_with_repository()
-        logger.info("TipiService initialized in Repository mode (Postgres)")
-        return
-
-    # Fallback temporário: enquanto a carga TIPI no Postgres não estiver completa.
-    app.state.tipi_service = TipiService()
-    logger.info(
-        "TipiService initialized in SQLite mode "
-        "(tipi.db - TIPI data not in Postgres yet)"
+def _build_cors_configuration() -> tuple[list[str], str | None]:
+    return _build_cors_configuration_impl(
+        settings.server.env,
+        settings.server.cors_allowed_origins,
+        settings.server.cors_allowed_origin_regex,
     )
 
 
-async def _init_nbs_service(app: FastAPI) -> None:
-    if settings.database.is_postgres:
-        try:
-            app.state.nbs_service = await NbsService.initializeNbsServiceWithPostgresRepository()
-            logger.info("NbsService initialized in Repository mode (Postgres)")
-            return
-        except Exception as e:
-            logger.warning(
-                "NbsService repository init failed, falling back to SQLite mode: %s",
-                e,
-            )
-    app.state.nbs_service = NbsService()
-    logger.info("NbsService initialized in SQLite mode (services.db)")
-
-
-async def _shutdown_resources(app: FastAPI) -> None:
-    logger.info("Shutting down...")
-
-    if hasattr(app.state, "db") and app.state.db:
-        try:
-            await app.state.db.close()
-        except Exception as e:
-            logger.warning("Error closing DatabaseAdapter: %s", e)
-
-    try:
-        await redis_cache.close()
-    except Exception as e:
-        logger.warning("Error closing Redis cache: %s", e)
-
-    if hasattr(app.state, "nbs_service") and app.state.nbs_service:
-        try:
-            await app.state.nbs_service.shutdownNbsServiceResources()
-        except Exception as e:
-            logger.warning("Error closing NbsService: %s", e)
-
-    if not getattr(app.state, "sqlmodel_enabled", False):
-        return
-
-    try:
-        from backend.infrastructure.db_engine import close_db
-
-        await close_db()
-        logger.info("SQLModel engine closed")
-    except Exception as e:
-        logger.warning("Error closing SQLModel engine: %s", e)
+_init_primary_database = _sync_lifecycle_async_call(_init_primary_database_impl)
+_init_sqlmodel_engine = _sync_lifecycle_async_call(_init_sqlmodel_engine_impl)
+_init_nesh_service = _sync_lifecycle_async_call(_init_nesh_service_impl)
+_init_cache_warmup = _sync_lifecycle_async_call(_init_cache_warmup_impl)
+_postgres_tipi_has_data = _sync_lifecycle_async_call(_postgres_tipi_has_data_impl)
+_init_tipi_service = _sync_lifecycle_async_call(_init_tipi_service_impl)
+_init_nbs_service = _sync_lifecycle_async_call(_init_nbs_service_impl)
+_shutdown_resources = _sync_lifecycle_async_call(_shutdown_resources_impl)
+_log_runtime_security_warnings = _sync_lifecycle_call(
+    _log_runtime_security_warnings_impl
+)
+_validate_dev_tenant_override_safety = _sync_lifecycle_call(
+    _validate_dev_tenant_override_safety_impl
+)
 
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
+    """Initializa e encerra recursos da aplicação de forma controlada."""
     project_root = _resolve_project_root()
     try:
         _record_release_metadata(app)
@@ -423,10 +206,7 @@ async def lifespan(app: FastAPI):
 
         logger.info("Initializing Glossary...")
         init_glossary(project_root)
-
-        # Check Frontend Build
         verify_frontend_build(project_root)
-
         yield
     finally:
         await _shutdown_resources(app)
@@ -434,48 +214,17 @@ async def lifespan(app: FastAPI):
 
 app = FastAPI(title="Nesh API", version="4.2", lifespan=lifespan)
 
-# --- Global Exception Handlers ---
 app.add_exception_handler(NeshError, cast(Any, nesh_exception_handler))
 app.add_exception_handler(Exception, cast(Any, generic_exception_handler))
 
-# --- Middleware ---
-# GZip (Architecture Improvement)
-# compresslevel=1 →  ~5x faster than level 6 with only ~10% larger output.
-# Critical: chapter responses are ~860KB; level 6 costs ~72ms, level 1 costs ~12ms.
 app.add_middleware(GZipMiddleware, minimum_size=1000, compresslevel=1)
-
-# Multi-tenant context middleware
 app.add_middleware(TenantMiddleware)
 
-# CORS Setup
-# Must be added after other middlewares so it wraps all responses,
-# including auth/tenant errors returned before route handlers.
-cors_origins = settings.server.cors_allowed_origins or [
-    "http://localhost:5173",
-    "http://127.0.0.1:5173",
-    "https://ysraestudos.github.io",
-]
-
-# Optional regex allows controlled preview domains (for example Cloudflare Pages)
-# while keeping explicit origin allowlists for production domains.
-cors_regex_parts: list[str] = []
-configured_cors_regex = (settings.server.cors_allowed_origin_regex or "").strip()
-if configured_cors_regex:
-    cors_regex_parts.append(f"(?:{configured_cors_regex})")
-
-# Dev convenience: allow local-network Vite hosts on :5173
-# without opening broad CORS in production.
-if settings.server.env == "development":
-    cors_regex_parts.append(
-        r"(?:^https?://(?:localhost|127\.0\.0\.1|\d{1,3}(?:\.\d{1,3}){3})(?::5173)?$)"
-    )
-
-cors_allow_origin_regex = "|".join(cors_regex_parts) or None
-
+_cors_allowed_origins, _cors_allow_origin_regex = _build_cors_configuration()
 app.add_middleware(
     CORSMiddleware,
-    allow_origins=cors_origins,
-    allow_origin_regex=cors_allow_origin_regex,
+    allow_origins=_cors_allowed_origins,
+    allow_origin_regex=_cors_allow_origin_regex,
     allow_credentials=True,
     allow_methods=["GET", "POST", "PATCH", "DELETE", "OPTIONS"],
     allow_headers=[
@@ -488,16 +237,15 @@ app.add_middleware(
         "Asaas-Access-Token",
         "X-Asaas-Access-Token",
     ],
-    expose_headers=[
-        "X-Request-Id",
-    ],
+    expose_headers=["X-Request-Id"],
 )
 
 
 @app.middleware("http")
 async def no_cache_html(request: Request, call_next):
+    """Aplica headers de segurança e evita cache em páginas HTML e docs."""
     path = request.url.path
-    if path in _DOCS_PATHS and not _should_expose_api_docs():
+    if path in {"/docs", "/redoc", "/openapi.json"} and not _should_expose_api_docs():
         response = JSONResponse(status_code=404, content={"detail": "Not Found"})
         _apply_security_headers(request, response)
         return response
@@ -512,41 +260,4 @@ async def no_cache_html(request: Request, call_next):
     return response
 
 
-# --- Routers ---
-# Prefixing to keep existing contract
-app.include_router(auth.router, prefix="/api", tags=["Auth"])
-app.include_router(search.router, prefix="/api", tags=["Search"])
-app.include_router(
-    tipi.router, prefix="/api/tipi", tags=["TIPI"]
-)  # Note: routes inside are /search, /chapters etc
-app.include_router(services.router, prefix="/api/services", tags=["Services"])
-app.include_router(database_download.router, prefix="/api/database", tags=["Database"])
-app.include_router(system.router, prefix="/api", tags=["System"])
-app.include_router(webhooks.router, prefix="/api/webhooks", tags=["Webhooks"])
-app.include_router(comments.router, prefix="/api", tags=["Comments"])
-app.include_router(profile.router, prefix="/api", tags=["Profile"])
-
-
-# --- Static Files / Frontend ---
-# Serving Frontend (Production Build)
-# Mounts the Vite build directory to serve the React App
-project_root = _resolve_project_root()
-static_dir = os.path.join(project_root, "client", "dist")
-
-if os.path.exists(static_dir):
-    app.mount("/", StaticFiles(directory=static_dir, html=True), name="static")
-else:
-    logger.warning(f"Frontend build not found at {static_dir}. Serving defaults.")
-
-    @app.get("/")
-    async def read_root():
-        return {
-            "message": (
-                "Nesh API running. Frontend not found. "
-                "Run 'npm run build' in client/ folder."
-            )
-        }
-
-    @app.head("/", include_in_schema=False)
-    async def read_root_head():
-        return Response(status_code=200)
+_configure_routes(app, _resolve_project_root(), logger)
